requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
babel==2.16.0
blinker==1.9.0
Bootstrap-Flask==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
docutils==0.21.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
libsql-client==0.3.1
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.57.1
packaging==24.2
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
sqlalchemy-libsql==0.1.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.0